Changed number to be a float instead of an int

+++ src/tree.rs
@@ -25,7 +25,7 @@ pub enum ExpType {
 
 #[derive(Debug, Clone)]
 pub enum Literal {
-    Number(i32),
+    Number(f32),
     Bool(bool),
     String(String),
 }
@@ -88,7 +88,7 @@ impl ExpType {
 
 impl Literal {
     pub fn new_number(token: &Token) -> Result<Literal, LunalaErrors> {
-        let value = token.access_lexeme().parse::<i32>();
+        let value = token.access_lexeme().parse::<f32>();
         match value {
             Ok(value) => Ok(Literal::Number(value)),
             Err(err) => Err(LunalaErrors::new(
